+++ src/world/import/SceneImporter.cpp
@@ -4,6 +4,15 @@ namespace world {
 
   SceneImporter::~SceneImporter() = default;
 
+  ImportOptionSchemas SceneImporter::editableSourceParameters(const QString&,
+                                                              const ImportOptions&) const {
+    return {};
+  }
+
+  bool SceneImporter::wrapDirectImportInSourceAsset() const {
+    return false;
+  }
+
   bool SceneImporter::configureImportedScene(Scene& scene, Element& importedRoot,
                                              const ImportOptions& options) const {
     (void)scene;